changes the auth setup to follow the nextauth.js example for clairity
