--- auth.config.ts
@@ -1,22 +1,0 @@
-import type { NextAuthConfig } from 'next-auth';
-import GitHub from 'next-auth/providers/github';
-
-export const authConfig = {
-    pages: {
-        signIn: '/login',
-    },
-    callbacks: {
-        authorized({ auth, request: { nextUrl } }) {
-            const isLoggedIn = !!auth?.user;
-            const isOnDashboard = nextUrl.pathname.startsWith('/dashboard');
-            if (isOnDashboard) {
-                if (isLoggedIn) return true;
-                return false; // Redirect unauthenticated users to login page
-            } else if (isLoggedIn) {
-                return Response.redirect(new URL('/dashboard', nextUrl));
-            }
-            return true;
-        },
-    },
-    providers: [GitHub], // Add providers with an empty array for now
-} satisfies NextAuthConfig;
