+++ auth.ts
@@ -1,6 +1,23 @@
 import NextAuth from "next-auth"
-import { authConfig } from "./auth.config"
+import GitHub from 'next-auth/providers/github';
+
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
-  ...authConfig
-})
+  pages: {
+    signIn: '/login',
+  },
+  callbacks: {
+    authorized({ auth, request: { nextUrl } }) {
+      const isLoggedIn = !!auth?.user;
+      const isOnDashboard = nextUrl.pathname.startsWith('/dashboard');
+      if (isOnDashboard) {
+        if (isLoggedIn) return true;
+        return false; // Redirect unauthenticated users to login page
+      } else if (isLoggedIn) {
+        return Response.redirect(new URL('/dashboard', nextUrl));
+      }
+      return true;
+    },
+  },
+  providers: [GitHub], // Add providers with an empty array for now
+});
