+++ middleware.ts
@@ -1,9 +1,7 @@
-import NextAuth from 'next-auth';
-import { authConfig } from './auth.config';
+export { auth as middleware } from "./auth"
 
-export default NextAuth(authConfig).auth;
 
 export const config = {
     // https://nextjs.org/docs/app/building-your-application/routing/middleware#matcher
-    matcher: ['/((?!api|_next/static|_next/image|.*\\.png$).*)'],
+    matcher: ["/((?!api|_next/static|_next/image|favicon.ico).*)"],
 };
